+++ C/src/kalman_filter.h
@@ -6,28 +6,28 @@
 class KalmanFilter
 {
 public:
-    KalmanFilter(const Eigen::Vector3d &initial_state);
+    KalmanFilter(const Eigen::Vector3f &initial_state);
 
     void predict();
-    void update(const Eigen::VectorXd &z);
-    std::pair<Eigen::VectorXd, Eigen::MatrixXd> filter_update(const Eigen::VectorXd *filtered_state_mean = nullptr,
-                                                              const Eigen::MatrixXd *filtered_state_covariance = nullptr,
-                                                              const Eigen::VectorXd *observation = nullptr);
+    void update(const Eigen::VectorXf &z);
+    std::pair<Eigen::VectorXf, Eigen::MatrixXf> filter_update(const Eigen::VectorXf *filtered_state_mean = nullptr,
+                                                              const Eigen::MatrixXf *filtered_state_covariance = nullptr,
+                                                              const Eigen::VectorXf *observation = nullptr);
 
     // Add getter methods for H and x_prior
-    const Eigen::MatrixXd &getH() const;
-    const Eigen::VectorXd &getXPrior() const;
-    const Eigen::VectorXd &getX() const;
+    const Eigen::MatrixXf &getH() const;
+    const Eigen::VectorXf &getXPrior() const;
+    const Eigen::VectorXf &getX() const;
 
 private:
-    Eigen::MatrixXd F;
-    Eigen::MatrixXd H;
-    Eigen::MatrixXd Q;
-    Eigen::MatrixXd R;
-    Eigen::VectorXd x;
-    Eigen::MatrixXd P;
-    Eigen::VectorXd x_prior;
-    Eigen::MatrixXd P_prior;
+    Eigen::MatrixXf F;
+    Eigen::MatrixXf H;
+    Eigen::MatrixXf Q;
+    Eigen::MatrixXf R;
+    Eigen::VectorXf x;
+    Eigen::MatrixXf P;
+    Eigen::VectorXf x_prior;
+    Eigen::MatrixXf P_prior;
 };
 
 #endif
